ArgenTech: rework engine-out airbraking
When we hit the QAssist altitude, we need to set the target airspeed
to a very low value. This allows the aircraft to nose-up and slow down.
This vastly improves the landings in all wind conditions, but especially
for light winds.

Additionally, I noticed that the FireEye has set a parameter that limits
the pitch angle to 1 degree during the QAssist phase. I have changed
that back to the default of 3 degrees, which allows for much better
airbraking.